Rename common to config
