+++ src/datarobot_genai/core/agents/__init__.py
@@ -22,7 +22,7 @@ This package provides:
   - llamaindex: DataRobotLiteLLM, create_pipeline_interactions_from_events
 """
 
-from ..mcp.common import MCPConfig
+from ..mcp.config import MCPConfig
 from .base import BaseAgent
 from .base import InvokeReturn
 from .base import UsageMetrics
