+++ src/datarobot_genai/crewai/__init__.py
@@ -4,7 +4,7 @@ Public API:
 - mcp_tools_context: Context manager returning available MCP tools for CrewAI.
 """
 
-from datarobot_genai.core.mcp.common import MCPConfig
+from datarobot_genai.core.mcp.config import MCPConfig
 
 from .agent import CrewAIAgent
 from .events import CrewAIRagasEventListener
