+++ src/datarobot_genai/llama_index/__init__.py
@@ -1,6 +1,6 @@
 """LlamaIndex utilities and helpers."""
 
-from datarobot_genai.core.mcp.common import MCPConfig
+from datarobot_genai.core.mcp.config import MCPConfig
 
 from .agent import DataRobotLiteLLM
 from .agent import LlamaIndexAgent
